+++ CHANGELOG.md
@@ -9,6 +9,11 @@
     * Add LsFilesOptions --format
 * Support git-annex 10.20230321 features
     * Deprecate getStatusAnx, isStatusAnx, statusAnx, StatusAnx, and StatusAnxOptions (documentation change only).
+* Make options the fourth generic function argument
+    * getFinds
+    * getLogs
+    * getShows
+    * getWhereis
 * The spawn error message includes environment variables only when env is supplied in ApiOptions.
 
 ## v2.2.0 released 2022-10-04